Combine 'util.tsx' and 'common.tsx' into a single 'Rivet.tsx'
