+++ src/components/index.tsx
@@ -2,4 +2,4 @@ export { default as Alert } from './Alert';
 export { default as Button } from './Button';
 export { default as Checkbox } from './Checkbox';
 export { default as Footer } from './Footer';
-export { Action } from './common';
+export { Action, Nav } from './Rivet';
